moved main part of the app in a separate component
